+++ src/main/java/com/budong/controller/NewsController.java
@@ -1,7 +1,8 @@
 package com.budong.controller;
 
+import java.text.SimpleDateFormat;
 import java.util.ArrayList;
-
+import java.util.Date;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -9,10 +10,13 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.servlet.ModelAndView;
-
 import com.budong.R;
 import com.budong.model.dto.NewsDTO;
 import com.budong.service.NewsServiceImpl;
+
+/* 
+ *	News Controller 
+ */
 
 @Controller
 public class NewsController {
@@ -20,6 +24,21 @@ public class NewsController {
 
     @Autowired
     private NewsServiceImpl newsService;
+    
+    @RequestMapping(R.mapping.khw_main)
+    public ModelAndView goToKhwIndexPage() {
+        Date today = new Date();
+        SimpleDateFormat date = new SimpleDateFormat("yyyyMMdd");
+        String param = date.format(today);
+
+        ModelAndView mav = new ModelAndView(R.path.khw_main);
+        String url = "https://land.naver.com/news/headline.nhn?bss_ymd=" + param;
+        ArrayList<NewsDTO> list = newsService.getTitle(url);
+
+        mav.addObject("list", list);
+        mav.addObject("date", param);
+        return mav;
+    }
 
     @RequestMapping(R.mapping.news_title)
     public ModelAndView goToTestPage(@RequestParam String param) {
